+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added configurable coloring for started todos with `todo.colors.started`. Fixes: https://github.com/vycdev/vscode-todo-plus-two/issues/59
 - Added a Todo files view toggle and `todo.file.view.showFinished` setting for hiding done and cancelled tasks. Fixes: https://github.com/vycdev/vscode-todo-plus-two/issues/34
 - Updated the embedded activity-bar view from open editor changes so source edits refresh without regenerating the embedded snapshot. Fixes: https://github.com/vycdev/vscode-todo-plus-two/issues/84
+- Added urgency coloring for `@due(...)` tags, configurable with `todo.colors.due.*` and `todo.due.soonDays`. Fixes: https://github.com/vycdev/vscode-todo-plus-two/issues/95
 
 ### Version 5.2.0